-*- All-new halo support. The iHaloRasterizer interface is gone; iHalo     objects are created instead. -*- Added halo support to MazeD (light setup dialog). Looks very nice.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@2272 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/renderer/software/scan.h
@@ -99,7 +99,7 @@ struct csScanSetup
   /// The fog pixel (R|G|B for true/hicolor) or index into palette of fog color
   int FogPix;
   /// Fog density
-  unsigned long FogDensity;
+  unsigned int FogDensity;
   /// The fog table for paletted (currently only 8-bit) modes
   unsigned char *Fog8;
